refactor: split pages and deduplicate sidebar components
