--- app/[locale]/(admin)/admin/_components/nav-main.tsx
@@ -1,47 +1,0 @@
-"use client"
-
-import { type LucideIcon } from "lucide-react"
-import { Link, usePathname } from "@/lib/navigation"
-
-import {
-  SidebarGroup,
-  SidebarGroupContent,
-  SidebarMenu,
-  SidebarMenuButton,
-  SidebarMenuItem,
-} from "@/components/ui/sidebar"
-
-export function NavMain({
-  items,
-}: {
-  items: {
-    title: string
-    url: string
-    icon?: LucideIcon
-  }[]
-}) {
-  const pathname = usePathname()
-
-  return (
-    <SidebarGroup className="px-2 py-2">
-      <SidebarGroupContent className="flex flex-col gap-1">
-        <SidebarMenu className="gap-1">
-          {items.map((item) => (
-            <SidebarMenuItem key={item.title}>
-              <SidebarMenuButton
-                tooltip={item.title}
-                isActive={pathname === item.url}
-                asChild
-              >
-                <Link href={item.url}>
-                  {item.icon && <item.icon />}
-                  <span>{item.title}</span>
-                </Link>
-              </SidebarMenuButton>
-            </SidebarMenuItem>
-          ))}
-        </SidebarMenu>
-      </SidebarGroupContent>
-    </SidebarGroup>
-  )
-}
