--- app/[locale]/(admin)/admin/_components/nav-secondary.tsx
@@ -1,49 +1,0 @@
-"use client"
-
-import { type LucideIcon } from "lucide-react"
-import { Link, usePathname } from "@/lib/navigation"
-
-import { cn } from "@/lib/utils"
-import {
-  SidebarGroup,
-  SidebarGroupContent,
-  SidebarMenu,
-  SidebarMenuButton,
-  SidebarMenuItem,
-} from "@/components/ui/sidebar"
-
-export function NavSecondary({
-  items,
-  ...props
-}: {
-  items: {
-    title: string
-    url: string
-    icon: LucideIcon
-  }[]
-} & React.ComponentPropsWithoutRef<typeof SidebarGroup>) {
-  const pathname = usePathname()
-
-  return (
-    <SidebarGroup {...props} className={cn("px-2 py-2", props.className)}>
-      <SidebarGroupContent>
-        <SidebarMenu className="gap-1">
-          {items.map((item) => (
-            <SidebarMenuItem key={item.title}>
-              <SidebarMenuButton
-                asChild
-                isActive={pathname === item.url}
-                size="sm"
-              >
-                <Link href={item.url}>
-                  <item.icon />
-                  <span>{item.title}</span>
-                </Link>
-              </SidebarMenuButton>
-            </SidebarMenuItem>
-          ))}
-        </SidebarMenu>
-      </SidebarGroupContent>
-    </SidebarGroup>
-  )
-}
