+++ app/[locale]/(admin)/admin/layout.tsx
@@ -3,8 +3,21 @@ import { requireSession } from "@/lib/auth-session"
 import { redirect } from "@/lib/navigation"
 import { SidebarInset, SidebarProvider } from "@/components/ui/sidebar"
 import { AppSidebar } from "./_components/app-sidebar"
-import { SiteHeader } from "./_components/site-header"
+import { SiteHeader } from "@/components/sidebar"
 import { ImpersonationBanner } from "@/components/impersonation-banner"
+
+const routeLabels: Record<string, string> = {
+  admin: "Admin",
+  users: "Users",
+  billing: "Billing",
+  analytics: "Analytics",
+  system: "System Health",
+  support: "Support",
+  audit: "Audit Log",
+  features: "Feature Flags",
+  settings: "Settings",
+  help: "Help",
+}
 
 export default async function AdminLayout({
   children,
@@ -41,7 +54,7 @@ export default async function AdminLayout({
       <AppSidebar variant="inset" user={sidebarUser} />
       <SidebarInset>
         <ImpersonationBanner session={session} />
-        <SiteHeader />
+        <SiteHeader routeLabels={routeLabels} />
         <div className="flex flex-1 flex-col">
           <div className="@container/main flex flex-1 flex-col gap-2">
             <div className="flex flex-col gap-4 py-4 md:gap-6 md:py-6">
