+++ app/[locale]/(admin)/admin/users/users-table.tsx
@@ -1,0 +1,264 @@
+"use client"
+
+import {
+  Table,
+  TableBody,
+  TableCell,
+  TableHead,
+  TableHeader,
+  TableRow,
+} from "@/components/ui/table"
+import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar"
+import { Badge } from "@/components/ui/badge"
+import { Button } from "@/components/ui/button"
+import { Input } from "@/components/ui/input"
+import {
+  Card,
+  CardContent,
+  CardDescription,
+  CardHeader,
+  CardTitle,
+} from "@/components/ui/card"
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuTrigger,
+} from "@/components/ui/dropdown-menu"
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog"
+import {
+  MoreHorizontal,
+  Ban,
+  UserX,
+  LogIn,
+  Loader2,
+  Shield,
+  Monitor,
+} from "lucide-react"
+import type { UserWithRole } from "./types"
+
+function getInitials(name: string | null) {
+  if (!name) return "U"
+  return name
+    .split(" ")
+    .map((n) => n[0])
+    .join("")
+    .toUpperCase()
+    .slice(0, 2)
+}
+
+export function UsersTable({
+  users,
+  isLoading,
+  searchQuery,
+  onSearchChange,
+  actionLoading,
+  currentUserId,
+  onBanUser,
+  onUnbanUser,
+  onImpersonate,
+  onOpenRoleDialog,
+  onOpenSessionsDialog,
+}: {
+  users: UserWithRole[]
+  isLoading: boolean
+  searchQuery: string
+  onSearchChange: (value: string) => void
+  actionLoading: string | null
+  currentUserId: string | undefined
+  onBanUser: (userId: string) => void
+  onUnbanUser: (userId: string) => void
+  onImpersonate: (userId: string) => void
+  onOpenRoleDialog: (user: UserWithRole) => void
+  onOpenSessionsDialog: (user: UserWithRole) => void
+}) {
+  return (
+    <Card>
+      <CardHeader>
+        <div className="flex items-center justify-between">
+          <div>
+            <CardTitle>Users</CardTitle>
+            <CardDescription>
+              A list of all registered users in the system
+            </CardDescription>
+          </div>
+          <Input
+            placeholder="Search by email..."
+            value={searchQuery}
+            onChange={(e) => onSearchChange(e.target.value)}
+            className="max-w-sm"
+          />
+        </div>
+      </CardHeader>
+      <CardContent>
+        {isLoading ? (
+          <div className="flex justify-center py-8">
+            <Loader2 className="h-8 w-8 animate-spin" />
+          </div>
+        ) : (
+          <Table>
+            <TableHeader>
+              <TableRow>
+                <TableHead>User</TableHead>
+                <TableHead>Role</TableHead>
+                <TableHead>Email Verified</TableHead>
+                <TableHead>Status</TableHead>
+                <TableHead>Created</TableHead>
+                <TableHead className="text-right">Actions</TableHead>
+              </TableRow>
+            </TableHeader>
+            <TableBody>
+              {users.length === 0 ? (
+                <TableRow>
+                  <TableCell
+                    colSpan={6}
+                    className="text-center text-muted-foreground"
+                  >
+                    No users found
+                  </TableCell>
+                </TableRow>
+              ) : (
+                users.map((user) => (
+                  <TableRow key={user.id}>
+                    <TableCell>
+                      <div className="flex items-center gap-3">
+                        <Avatar className="h-8 w-8">
+                          <AvatarImage
+                            src={user.image || ""}
+                            alt={user.name || ""}
+                          />
+                          <AvatarFallback>
+                            {getInitials(user.name)}
+                          </AvatarFallback>
+                        </Avatar>
+                        <div>
+                          <p className="font-medium">
+                            {user.name || "No name"}
+                          </p>
+                          <p className="text-sm text-muted-foreground">
+                            {user.email}
+                          </p>
+                        </div>
+                      </div>
+                    </TableCell>
+                    <TableCell>
+                      {user.role === "admin" ? (
+                        <Badge className="bg-purple-100 text-purple-800 dark:bg-purple-900 dark:text-purple-200">
+                          <Shield className="mr-1 h-3 w-3" />
+                          Admin
+                        </Badge>
+                      ) : (
+                        <Badge variant="outline">User</Badge>
+                      )}
+                    </TableCell>
+                    <TableCell>
+                      {user.emailVerified ? (
+                        <Badge variant="secondary">Verified</Badge>
+                      ) : (
+                        <Badge variant="outline">Unverified</Badge>
+                      )}
+                    </TableCell>
+                    <TableCell>
+                      {user.banned ? (
+                        <Badge variant="destructive">Banned</Badge>
+                      ) : (
+                        <Badge variant="outline">Active</Badge>
+                      )}
+                    </TableCell>
+                    <TableCell>
+                      {new Date(user.createdAt).toLocaleDateString()}
+                    </TableCell>
+                    <TableCell className="text-right">
+                      <DropdownMenu>
+                        <DropdownMenuTrigger asChild>
+                          <Button
+                            variant="ghost"
+                            size="sm"
+                            disabled={actionLoading === user.id}
+                          >
+                            {actionLoading === user.id ? (
+                              <Loader2 className="h-4 w-4 animate-spin" />
+                            ) : (
+                              <MoreHorizontal className="h-4 w-4" />
+                            )}
+                          </Button>
+                        </DropdownMenuTrigger>
+                        <DropdownMenuContent align="end">
+                          <DropdownMenuItem
+                            onClick={() => onOpenRoleDialog(user)}
+                          >
+                            <Shield className="mr-2 h-4 w-4" />
+                            Change Role
+                          </DropdownMenuItem>
+                          <DropdownMenuItem
+                            onClick={() => onOpenSessionsDialog(user)}
+                          >
+                            <Monitor className="mr-2 h-4 w-4" />
+                            Sessions
+                          </DropdownMenuItem>
+                          <DropdownMenuItem
+                            onClick={() => onImpersonate(user.id)}
+                            disabled={user.id === currentUserId}
+                          >
+                            <LogIn className="mr-2 h-4 w-4" />
+                            Impersonate
+                          </DropdownMenuItem>
+                          {user.banned ? (
+                            <DropdownMenuItem
+                              onClick={() => onUnbanUser(user.id)}
+                            >
+                              <UserX className="mr-2 h-4 w-4" />
+                              Unban User
+                            </DropdownMenuItem>
+                          ) : (
+                            <AlertDialog>
+                              <AlertDialogTrigger asChild>
+                                <DropdownMenuItem
+                                  onSelect={(e) => e.preventDefault()}
+                                >
+                                  <Ban className="mr-2 h-4 w-4" />
+                                  Ban User
+                                </DropdownMenuItem>
+                              </AlertDialogTrigger>
+                              <AlertDialogContent>
+                                <AlertDialogHeader>
+                                  <AlertDialogTitle>Ban User</AlertDialogTitle>
+                                  <AlertDialogDescription>
+                                    Are you sure you want to ban {user.email}?
+                                    This will prevent them from signing in.
+                                  </AlertDialogDescription>
+                                </AlertDialogHeader>
+                                <AlertDialogFooter>
+                                  <AlertDialogCancel>Cancel</AlertDialogCancel>
+                                  <AlertDialogAction
+                                    onClick={() => onBanUser(user.id)}
+                                  >
+                                    Ban User
+                                  </AlertDialogAction>
+                                </AlertDialogFooter>
+                              </AlertDialogContent>
+                            </AlertDialog>
+                          )}
+                        </DropdownMenuContent>
+                      </DropdownMenu>
+                    </TableCell>
+                  </TableRow>
+                ))
+              )}
+            </TableBody>
+          </Table>
+        )}
+      </CardContent>
+    </Card>
+  )
+}
