+++ app/[locale]/(dashboard)/_components/app-sidebar.tsx
@@ -11,9 +11,7 @@ import {
 } from "lucide-react"
 
 import { Link } from "@/lib/navigation"
-import { NavMain } from "./nav-main"
-import { NavSecondary } from "./nav-secondary"
-import { NavUser } from "./nav-user"
+import { NavMain, NavSecondary, NavUser } from "@/components/sidebar"
 import {
   Sidebar,
   SidebarContent,
@@ -88,7 +86,7 @@ export function AppSidebar({
         <NavSecondary items={data.navSecondary} className="mt-auto" />
       </SidebarContent>
       <SidebarFooter>
-        <NavUser user={user} />
+        <NavUser user={user} billingPath="/dashboard" />
       </SidebarFooter>
     </Sidebar>
   )
