+++ app/[locale]/not-found.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import { useTranslations } from "next-intl"
 
 export default function NotFound() {
