+++ components/sidebar/index.ts
@@ -1,0 +1,4 @@
+export { NavMain } from "./nav-main"
+export { NavSecondary } from "./nav-secondary"
+export { NavUser } from "./nav-user"
+export { SiteHeader } from "./site-header"
